refactor(exception): custom exception들을 관리하는 추상클래스 작성
