+++ src/main/java/gift/global/exception/custrom/DBException.java
@@ -2,21 +2,8 @@ package gift.global.exception.custrom;
 
 import gift.global.exception.ErrorCode;
 
-public class DBException extends RuntimeException {
-    private final ErrorCode errorCode;
-    private final String details;
-
+public class DBException extends CustomException {
     public DBException(ErrorCode errorCode, String details) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-        this.details = details;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-
-    public String getDetails() {
-        return details;
+        super(errorCode, details);
     }
 }
