+++ src/main/java/gift/global/exception/custrom/LoginException.java
@@ -2,22 +2,8 @@ package gift.global.exception.custrom;
 
 import gift.global.exception.ErrorCode;
 
-public class LoginException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-    private final String details;
-
+public class LoginException extends CustomException {
     public LoginException(ErrorCode errorCode, String details) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-        this.details = details;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-
-    public String getDetails() {
-        return details;
+        super(errorCode, details);
     }
 }
